[EN SD] Update regexes for CORS Origin format
According to the Chrome Developer Tools, these headers are sent in with
CORS requests.

```
Origin: http://localhost:8080 (when testing locally)
```
```
Origin: https://beta.turbovote.org (on the beta site)
```

We've corrected the regexes to account for these changes.

Pedestal uses the entire value of the Origin header to make the
comparison.